event_server/requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
langgraph_sdk>=0.1.25,<0.2.0
langchain_core>=0.2.11,<0.3.0
qdrant-client
